requirements.txt
paho-mqtt~=1.6.1

psutil~=5.9.5

py-cpuinfo~=9.0.0